Free_Data_API-feat-microservice-refactor/financial_data_framework/requirements.txt
# 核心依賴
yfinance
duckdb
PyYAML
requests
pandas
